+++ src/integration/agents/src/main/java/se/tink/backend/aggregation/agents/nxgen/serviceproviders/creditcards/sebkort/fetcher/entity/TransactionEntity.java
@@ -2,6 +2,7 @@ package se.tink.backend.aggregation.agents.nxgen.serviceproviders.creditcards.se
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import java.math.BigDecimal;
 import java.util.Date;
 import se.tink.backend.aggregation.agents.nxgen.serviceproviders.creditcards.sebkort.SebKortConstants;
 import se.tink.backend.aggregation.annotations.JsonObject;
@@ -19,9 +20,9 @@ public class TransactionEntity {
     private Date date;
 
     private String countryCode;
-    private double amount;
+    private BigDecimal amount;
     private String currencyCode;
-    private double billingAmount;
+    private BigDecimal billingAmount;
     private String billingCurrencyCode;
     private String type;
 
@@ -60,7 +61,7 @@ public class TransactionEntity {
         return countryCode;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
@@ -68,7 +69,7 @@ public class TransactionEntity {
         return currencyCode;
     }
 
-    public double getBillingAmount() {
+    public BigDecimal getBillingAmount() {
         return billingAmount;
     }
